requirements.txt
streamlit
pandas
pillow
google-generativeai
python-dotenv